Refactor: Introduce redux starter kit
This remove redux hooks usage in the library and instead stick with the
use of normal redux workflow. It introduce creatorSlice Factory which
will be used to create slices easly without too much boilerplate codes

Also the logics in the slice factory have been tested and make sure the
reducers created will work as specified.

close #1